+++ py_ce_forms_api/form/form.py
@@ -42,6 +42,12 @@ class Form(FormCore):
     def get_node_form(self, field:str) -> Form:
         return self._get_stored_form("nodes", field)
 
+    def get_nodes_forms(self) -> list[Form]:
+        nodes = self.form.get('nodes')
+        if nodes is None:
+            return []
+        return [Form(node) for node in nodes.values()]
+
     def get_root(self) -> str:
         return self.form["root"]
     
@@ -67,6 +73,9 @@ class Form(FormCore):
     def apply_on_blocks(self, func: function):
         for block in self.form["content"].values():
             func(FormBlock(self, block))
+    
+    def get_blocks(self):
+        return [FormBlock(self, block) for block in self.form["content"].values()]
     
     def _get_stored_form(self, part: str, field:str) -> Form:
         if self.form.get(part) is None or self.form[part].get(field) is None: