+++ py_ce_forms_api/processing/processing.py
@@ -1,9 +1,12 @@
 import os
-from fastapi import FastAPI, APIRouter
+
+from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
+from fastapi import Depends, FastAPI, APIRouter, HTTPException, status
 from ..client import CeFormsClient
 from .processing_tasks import ProcessingTasks
 
 app = FastAPI()
+security = HTTPBearer()
 
 class Processing(ProcessingTasks):
     """
@@ -14,9 +17,16 @@ class Processing(ProcessingTasks):
     def __init__(self, client: CeFormsClient, func) -> None:
         ProcessingTasks.__init__(self, client, func)
         self.server = "localhost"
-        self.port = os.environ.get("CE_FORMS_TASK_PORT")       
+        self.port = os.environ.get("CE_FORMS_TASK_PORT")
+        self.token = os.environ.get("CE_FORMS_TASK_TOKEN")
+        
+        if self.token is None:
+            print("[Processing] There is no defined token to protect from externals accesses")
+        
         self.app = app           
-        self.router = APIRouter()  
+        self.router = APIRouter(
+            dependencies= [Depends(self._verify_token)] if self.token is not None else []
+        )  
         self.router.add_api_route("/", self.self, methods=["GET"])    
         self.router.add_api_route("/processing/{pid}", self.__do_processing, methods=["GET"])
         self.router.add_api_route("/cancel/{pid}", self.__cancel, methods=["GET"])
@@ -31,6 +41,15 @@ class Processing(ProcessingTasks):
             
     def __cancel(self, pid: str):        
         return self.cancel(pid)       
+    
+    def _verify_token(self, credentials: HTTPAuthorizationCredentials = Depends(security)):                        
+        token = credentials.credentials
+        if token != self.token:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Invalid or missing Bearer token",
+            )
+        return token
     
     def self(self):
         return self.tasks.status()